Saves size phantom masks as 32-bit instead of 64

+++ notes/2018-09-05-retune-parameters/phantoms/make_different_size.py
@@ -5,4 +5,5 @@ from phantom_funcs import make_bg_phantom, make_orth_phantom
 for r in [4, 8, 12, 16]:
     phant, mask = make_orth_phantom(2, 9, r)
     imsave('./different_size/phants/x_phantom_nfib9_r{}.tif'.format(r), phant)
-    imsave('./different_size/masks/x_phantom_nfib9_r{}.tif'.format(r), mask)
+    imsave('./different_size/masks/x_phantom_nfib9_r{}.tif'.format(r),
+           mask, dtype=np.float32)
